+++ Client/Includes/ClientEntityPool.hpp
@@ -5,7 +5,7 @@
 #ifndef R_TYPE_CLIENTENTITYPOOL_HPP
 #define R_TYPE_CLIENTENTITYPOOL_HPP
 
-#include <EntityPool.hpp>
+#include <EntityPool/EntityPool.hpp>
 
 class ClientEntityPool : public EntityPool {
 public :
