+++ Client/src/RTypeGameContext.cpp
@@ -3,8 +3,8 @@
 //
 
 #include <Timer.hpp>
-#include <ManagedExternalInstance.hpp>
-#include <ExternalClassFactoryLoader.hpp>
+#include <LibraryLoader/ManagedExternalInstance.hpp>
+#include <LibraryLoader/ExternalClassFactoryLoader.hpp>
 #include "RTypeGameContext.hpp"
 
 RTypeGameContext::RTypeGameContext() : _pool() {
