+++ Shared/EntityPool/EntityPool.cpp
@@ -5,8 +5,8 @@
 #include <vec2.hpp>
 #include <DrawableEntity.hpp>
 #include <iostream>
-#include "EntityPool.hpp"
-#include "../LibraryLoader/ManagedExternalInstance.hpp"
+#include "EntityPool/EntityPool.hpp"
+#include "LibraryLoader/ManagedExternalInstance.hpp"
 
 void EntityPool::AddEntity(ManagedExternalInstance<Entity> &entity) {
     _pool.push_back(entity);
