+++ Shared/EventDispatcher/EventManager.cpp
@@ -2,7 +2,7 @@
 // Created by aguado_e on 11/21/16.
 //
 
-#include <EventManager.hpp>
+#include <EventDispatcher/EventManager.hpp>
 
 RType::EventManager::EventManager() {}
