+++ Shared/Include/LibraryLoader/ExternalClassFactory.hpp
