+++ Shared/Include/LibraryLoader/ManagedExternalInstance.hpp
