+++ Shared/Include/PartitionSystem/EntityPartitionBuilder.hpp
@@ -5,7 +5,7 @@
 #ifndef R_TYPE_ENTITYPARTITIONBUILDER_HPP
 #define R_TYPE_ENTITYPARTITIONBUILDER_HPP
 
-#include "PartitionSegmentBuilder.hpp"
+#include "PartitionSystem/PartitionSegmentBuilder.hpp"
 #include "EntityPartition.hpp"
 
 class EntityPartitionBuilder {
