+++ Shared/Include/PartitionSystem/Tween/Curve/ITweeningCurve.hpp
