+++ Shared/Include/PartitionSystem/Tween/Curve/SmoothStepCurve.hpp
@@ -6,7 +6,7 @@
 #define R_TYPE_SMOOTHSTEPCURVE_HPP
 
 
-#include "ITweeningCurve.hpp"
+#include "PartitionSystem/Tween/Curve/ITweeningCurve.hpp"
 
 class SmoothStepCurve : public ITweeningCurve {
 public:
