+++ Shared/LibraryLoader/CMakeLists.txt
@@ -3,19 +3,19 @@ cmake_minimum_required(VERSION 2.6.2)
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
 
 set(LIBRARY_LOADER_FILES
-        ExternalClassFactoryLoader.hpp
-        ManagedExternalInstance.hpp
-        ExternalClassFactory.hpp)
+        ../Include/LibraryLoader/ExternalClassFactoryLoader.hpp
+        ../Include/LibraryLoader/ManagedExternalInstance.hpp
+        ../Include/LibraryLoader/ExternalClassFactory.hpp)
 include_directories(../Include)
 
 IF (MSVC)
     set(LIBRARY_LOADER_FILES ${LIBRARY_LOADER_FILES}
-            InternalWindowsLibraryLoader.hpp
+            ../Include/LibraryLoader/InternalWindowsLibraryLoader.hpp
             InternalWindowsLibraryLoader.cpp)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
 ELSE()
     set(LIBRARY_LOADER_FILES ${LIBRARY_LOADER_FILES}
-            InternalLinuxLibraryLoader.hpp
+            ../Include/LibraryLoader/InternalLinuxLibraryLoader.hpp
             InternalLinuxLibraryLoader.cpp)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11 -Wall -Wextra -Werror -Wuninitialized -Winit-self -Weffc++ -Wno-unused-parameter")
 ENDIF()
