+++ SharedLibs/Entities/Monsters/DummyMonster/DummyMonster.hpp
@@ -5,12 +5,12 @@
 #ifndef R_TYPE_DUMMYMONSTER_HPP
 #define R_TYPE_DUMMYMONSTER_HPP
 
-#include "EntityPartition.hpp"
+#include "PartitionSystem/EntityPartition.hpp"
 #include <Timer.hpp>
 #include <Base.h>
 #include <IDrawable.hpp>
 #include <DrawableEntity.hpp>
-#include "PartitionSegmentBuilder.hpp"
+#include "PartitionSystem/PartitionSegmentBuilder.hpp"
 
 class DummyMonster : public Entity {
 protected:
